+++ changelog.md
@@ -13,6 +13,8 @@ SciReactUI Changelog
 - Navbar, NavLink and FooterLink will use routing library for links if provided with linkComponent and to props.
 - Navbar uses slots for positioning elements. Breaking change: elements must now use rightSlot for positioning to the far right.
 - User can take additional menu items through the menuItems prop.
+- Footer uses slots for positioning elements. Breaking change: elements must now use rightSlot for positioning to the far right.
+
 
 [v0.1.0] - 2025-04-10
 ---------------------